moves common funcs to utils
