+++ src/colors/histogram/histogram.ts
@@ -1,4 +1,5 @@
-import {ColorChannels, PaletteAlgorithm, Pixel} from '../colorPalette';
+import { ColorChannels, PaletteAlgorithm, Pixel } from '../colorPalette';
+import { getAverageColors, getPixelHashKey } from '../utils';
 
 type PixelHash = Record<string, Pixel[]>;
 type Sector = [string, Pixel[]];
@@ -22,13 +23,14 @@ class Histogram implements PaletteAlgorithm {
     const sectorOrder: Sector[] = Object.entries(pixelHash).sort((a, b) => {
       return b[1].length - a[1].length;
     });
-    const paletteSize = sectorOrder.length > this.paletteSize
-      ? this.paletteSize
-      : sectorOrder.length;
+    const paletteSize =
+      sectorOrder.length > this.paletteSize
+        ? this.paletteSize
+        : sectorOrder.length;
 
     for (let i = 0; i < paletteSize; i++) {
       const sector = sectorOrder[i][1];
-      const averageColor = this.getAverageColors(sector);
+      const averageColor = getAverageColors(sector);
       palette.push(averageColor);
     }
 
@@ -43,43 +45,16 @@ class Histogram implements PaletteAlgorithm {
       const g = greenChannel[i];
       const b = blueChannel[i];
 
-      const key = this.getPixelHashKey({r, g, b});
+      const key = getPixelHashKey({ r, g, b }, this.dimensions);
 
       if (key in pixelHash) {
-        pixelHash[key].push({r, g, b});
+        pixelHash[key].push({ r, g, b });
       } else {
-        pixelHash[key] = [{r, g, b}];
+        pixelHash[key] = [{ r, g, b }];
       }
     }
 
     return pixelHash;
-  }
-
-  getPixelHashKey(pixel: Pixel) {
-    const redSector = Math.floor(pixel.r / this.dimensions);
-    const greenSector = Math.floor(pixel.g / this.dimensions);
-    const blueSector = Math.floor(pixel.b / this.dimensions);
-
-    return `${redSector}_${greenSector}_${blueSector}`;
-  }
-
-  getAverageColors(groupedPixels: Pixel[]): Pixel {
-    const totalPixels = groupedPixels.length;
-    const redSum = groupedPixels.reduce((sum, pixel) => {
-      return sum + pixel.r;
-    }, 0);
-    const greenSum = groupedPixels.reduce((sum, pixel) => {
-      return sum + pixel.g;
-    }, 0);
-    const blueSum = groupedPixels.reduce((sum, pixel) => {
-      return sum + pixel.b;
-    }, 0);
-
-    return {
-      r: Math.round(redSum / totalPixels),
-      g: Math.round(greenSum / totalPixels),
-      b: Math.round(blueSum / totalPixels),
-    };
   }
 }
